requirements.txt
blinker==1.9.0
certifi==2025.10.5
charset-normalizer==3.4.4
clarifai==10.8.2
clarifai-grpc==11.6.1
click==8.3.0
colorama==0.4.6
contextlib2==21.6.0
Flask==3.1.2
flask-cors==6.0.1
googleapis-common-protos==1.70.0
grpcio==1.75.1
gunicorn==23.0.0
idna==3.11
inquirerpy==0.3.4
itsdangerous==2.2.0
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
numpy
packaging==25.0
perf_analyzer==0.1.0
pfzy==0.3.4
pillow==12.0.0
prompt_toolkit==3.0.52
protobuf==5.27.3
Pygments==2.19.2
python-dotenv==1.1.1
python-rapidjson==1.21
PyYAML==6.0.3
requests==2.32.5
rich==14.2.0
schema==0.7.5
tabulate==0.9.0
tqdm==4.67.1
tritonclient==2.61.0
typing_extensions==4.15.0
urllib3==2.5.0
wcwidth==0.2.14
Werkzeug==3.1.3
